Improve vector (de-)serialization.
std::vector's data chunk is always a contiguous chunk of memory and can
therefore be serialized and deserialized in one go.

Special treatment of `std::vector<bool>` is required because direct
access to data() is forbidden in that case. Using the old implementation
here.

Also declaring the functions inline to avoid duplicate symbol errors.