Introduce internal option to disable threading.
Added the `enable_threading` flag to control whether parallel
computations should be done in separate threads or right in the context
of the calling thread.

It is not exported as a command line argument as it does not make much
sense there. But it is helpful when integrating ranger into another
application that already has its own thread management or when
creating separate threads per prediction introduce too much overhead.