Change Forest::data from Data* to std::unique_ptr<Data>
Refactor calling and using code to accomidate this change. Ultimatly, this means Forest now owns data, which it appeared to be doing anyway. This should help avoid leaks (e.g. data loaded from file was never deleted). Change Tree's Data* to const Data* to emphasize this is a non-owning pointer.