Allow locally scoped functions in lhs of formula
Add an environment parameter to `parse.formula` to allow locally scoped
functions in the left hand side of the formulas.

This commit affects the formula interface for `ranger::ranger` such that
calls of the form

    ranger(myLocalFunction(responseVar) ~ ., data = myData)

are possible evern if `myLocalFunction` is not part of the global
environment. See the corresponding issue for details.

Fixes #373