Add `verbose` option to `csrf()`
`csrf()` can be computationally demanding, and having an idea of the progress would be really beneficial. While there are many ways to accomplish this (e.g., progress bars, etc.), this approach does not require any changes to the core `csrf()` code or adds any dependencies. The default here is `FALSE`.